+++ docs/changelog.md
@@ -9,6 +9,7 @@ Date of release: *n/a*
 - Fixed the pyperf requirement which was originally pinning pyperf to 2.0.0 by accident.
 - Added pre-checks for both pyperf arguments and --format-config configuration so issues
   are caught earlier.
+- The parse task now supports installations of Black as old as 19.3b0!
 
 ## 21.7.dev2